% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ParaGraph.R
\name{paragraph}
\alias{paragraph}
\title{This package takes a quantitative and a qualitative variable and determine if they meet the assumptions for a one-way anova. If so, an anova is run and results, along with a post-hoc tukey test are shown. If data does not meet parametric assumptions than a kruskall wallis and subsequent dunn test are displayed. Data can be graphed producing a bar graph depicting the mean and SEM of "x" with the bars coloured by "y". Compact letters are also displayed above each bar.}
\usage{
paragraph(
  df,
  x,
  y,
  remove_outliers = "FALSE",
  post_hoc = "TRUE",
  graph = "FALSE"
)
}
\arguments{
\item{df}{A data frame.}

\item{x}{A numeric vector. The numerical variable name or column number.}

\item{y}{A character vector. The categorical variable name or column number.}

\item{remove_outliers}{A string. Optional, either "TRUE" or "FALSE". Defaults to "FALSE".}

\item{post_hoc}{A string. Determines if a post-hoc analysis is run. Optional, either "TRUE" or "FALSE". Defaults to "TRUE".}

\item{graph}{Determines if the output is a graph or the data frame. Optional, either "TRUE" or "FALSE". Defaults to "FALSE".}
}
\value{
A bar graph and/or analysis results.
}
\description{
This package takes a quantitative and a qualitative variable and determine if they meet the assumptions for a one-way anova. If so, an anova is run and results, along with a post-hoc tukey test are shown. If data does not meet parametric assumptions than a kruskall wallis and subsequent dunn test are displayed. Data can be graphed producing a bar graph depicting the mean and SEM of "x" with the bars coloured by "y". Compact letters are also displayed above each bar.
}
\examples{
paragraph(iris, 1, 5, graph = "TRUE")

paragraph(iris,"Sepal.Length", "Species", remove_outliers = "TRUE", graph = "FALSE")
}
